+++ test/enum_test.exs
@@ -356,11 +356,4 @@ defmodule A.EnumTest do
     assert 0 = A.Vector.new() |> A.Enum.max(&>=/2, fn -> 0 end)
     assert 0 = MapSet.new() |> A.Enum.max(&>=/2, fn -> 0 end)
   end
-
-  test "A.Enum.sort_uniq/1" do
-    assert [] = A.Enum.sort_uniq([])
-    assert [1, 2, 3] = A.Enum.sort_uniq([3, 3, 3, 2, 2, 1])
-    assert [1, 2, 3, 4, 5] = A.Enum.sort_uniq(5..1)
-    assert [1.0, 1, 2.0, 2] = A.Enum.sort_uniq([2, 1, 1.0, 2.0])
-  end
 end